+++ Godeps/_workspace/src/github.com/elastic/libbeat/outputs/elasticsearch/bulkapi.go
@@ -27,6 +27,10 @@ type bulkMetaIndex struct {
 	DocType string `json:"_type"`
 }
 
+type BulkResult struct {
+	Items []json.RawMessage `json:"items"`
+}
+
 func (r *bulkRequest) Send(meta, obj interface{}) error {
 	var err error
 
@@ -40,19 +44,20 @@ func (r *bulkRequest) Send(meta, obj interface{}) error {
 	return err
 }
 
-func (r *bulkRequest) Flush() (*QueryResult, error) {
+func (r *bulkRequest) Flush() (int, *BulkResult, error) {
 	if r.buf.Len() == 0 {
 		logp.Debug("elasticsearch", "Empty channel. Wait for more data.")
-		return nil, nil
+		return 0, nil, nil
 	}
 
-	resp, err := r.es.sendBulkRequest("POST", r.path, r.params, &r.buf)
+	status, resp, err := r.es.sendBulkRequest("POST", r.path, r.params, &r.buf)
 	if err != nil {
-		return nil, err
+		return status, nil, err
 	}
 	r.buf.Truncate(0)
 
-	return readQueryResult(resp)
+	result, err := readBulkResult(resp)
+	return status, result, err
 }
 
 // Bulk performs many index/delete operations in a single API call.
@@ -90,7 +95,7 @@ func (conn *Connection) BulkWith(
 		return nil, nil
 	}
 
-	resp, err := conn.sendBulkRequest("POST", path, params, &buf)
+	_, resp, err := conn.sendBulkRequest("POST", path, params, &buf)
 	if err != nil {
 		return nil, err
 	}
@@ -120,7 +125,7 @@ func (conn *Connection) sendBulkRequest(
 	method, path string,
 	params map[string]string,
 	buf *bytes.Buffer,
-) ([]byte, error) {
+) (int, []byte, error) {
 	url := makeURL(conn.URL, path, params)
 	logp.Debug("elasticsearch", "Sending bulk request to %s", url)
 
@@ -153,4 +158,17 @@ func bulkEncode(metaBuilder MetaBuilder, body []interface{}) bytes.Buffer {
 		}
 	}
 	return buf
+}
+
+func readBulkResult(obj []byte) (*BulkResult, error) {
+	if obj == nil {
+		return nil, nil
+	}
+
+	var result BulkResult
+	err := json.Unmarshal(obj, &result)
+	if err != nil {
+		return nil, err
+	}
+	return &result, nil
 }