+++ Godeps/_workspace/src/github.com/elastic/libbeat/publisher/preprocess.go
@@ -66,7 +66,7 @@ func (p *preprocessor) onMessage(m message) {
 	// return if no event is left
 	if len(ignore) == len(events) {
 		debug("no event left, complete send")
-		outputs.SignalCompleted(m.signal)
+		outputs.SignalCompleted(m.context.signal)
 		return
 	}
 
@@ -84,29 +84,29 @@ func (p *preprocessor) onMessage(m message) {
 
 	if publisher.disabled {
 		debug("publisher disabled")
-		outputs.SignalCompleted(m.signal)
+		outputs.SignalCompleted(m.context.signal)
 		return
 	}
 
 	debug("preprocessor forward")
 	if single {
-		p.handler.onMessage(message{signal: m.signal, event: events[0]})
+		p.handler.onMessage(message{context: m.context, event: events[0]})
 	} else {
-		p.handler.onMessage(message{signal: m.signal, events: events})
+		p.handler.onMessage(message{context: m.context, events: events})
 	}
 }
 
 // filterEvent validates an event for common required fields with types.
 // If event is to be filtered out the reason is returned as error.
 func filterEvent(event common.MapStr) error {
-	ts, ok := event["timestamp"]
+	ts, ok := event["@timestamp"]
 	if !ok {
-		return errors.New("Missing 'timestamp' field from event")
+		return errors.New("Missing '@timestamp' field from event")
 	}
 
 	_, ok = ts.(common.Time)
 	if !ok {
-		return errors.New("Invalid 'timestamp' field from event.")
+		return errors.New("Invalid '@timestamp' field from event.")
 	}
 
 	err := event.EnsureCountField()