+++ CHANGELOG.md
@@ -2,9 +2,21 @@
 All notable changes to this project will be documented in this file based on the
 [Keep a Changelog](http://keepachangelog.com/) Standard.
 
-## [Unreleased](https://github.com/elastic/packetbeat/compare/1.0.0-beta4...HEAD)
+## [Unreleased](https://github.com/elastic/libbeat/compare/1.0.0-rc1...HEAD)
 
 ### Backward Compatibility Breaks
+
+### Bugfixes
+- Packetbeat will now exit if a configuration error is detected. #357
+
+### Added
+
+### Deprecated
+
+## [1.0.0-rc1](https://github.com/elastic/libbeat/compare/1.0.0-beta4...1.0.0-rc1) 2015-11-04
+
+### Backward Compatibility Breaks
+- Rename timestamp field with @timestamp. #343
 
 ### Bugfixes
 - Close file descriptors used to monitor processes. #337
@@ -14,12 +26,12 @@ All notable changes to this project will be documented in this file based on the
 
 ### Deprecated
 
-## [1.0.0-beta4] - 2015-10-21
+## [1.0.0-beta4](https://github.com/elastic/libbeat/compare/1.0.0-beta3...1.0.0-beta4) 2015-10-21
 
 ### Backward Compatibility Breaks
-- renamed http module config file option 'strip_authorization' to 'redact_authorization'
-- save_topology is set to false by default
-- rename elasticsearch index to [packetbeat-]YYYY.MM.DD
+- Renamed http module config file option 'strip_authorization' to 'redact_authorization'
+- Save_topology is set to false by default
+- Rename elasticsearch index to [packetbeat-]YYYY.MM.DD
 
 ### Bugfixes
 - Support for lower-case header names when redacting http authorization headers